pyproject.toml
[tool.poetry]
name = "tenjin"
version = "0.1.0"
description = ""
authors = ["Tim Warren <hello@timwarren.dev>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
openai = "0.27.0"
langchain = "0.0.99"
gradio = "^3.16.2"
toml = "^0.10.2"
pinecone-client = "^2.1.0"
fitz = "^0.0.1.dev2"
pymupdf = "^1.21.1"
fastapi = "^0.89.1"
uvicorn = "^0.20.0"
pydantic = "^1.10.4"
google-api-python-client = "^2.77.0"
boto3 = "^1.26.66"
slack-sdk = "^3.19.5"
wolframalpha = "^5.0.0"
python-dotenv = "^0.21.1"
tiktoken = "^0.2.0"
faiss-cpu = "^1.7.3"

[tool.poetry.group.dev.dependencies]
black = "^22.12.0"
pytest = "^7.2.1"
pytest-watch = "^4.2.0"

[virtualenvs]
create = true
in-project = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
server = "tenjin:serve"
